+++ doc/Changelog.rst
@@ -10,6 +10,10 @@ Changelog
 * Fix bug kraken sequential when no reads are classified
 * Fix FileFactory to make it much faster (speed up pipeline initialisation). This should also fix a recurrent issue with read tag (see https://github.com/sequana/sequana/issues/526
 
+* NEWS:
+
+    * implement  go term enrichment
+
 
 0.8.6
 -----